Add VALIDATE_BCN option
